legal-rag-app/requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
numpy
openai
langchain
langchain-huggingface
sentence-transformers
rank-bm25
# PyMuPDF is not needed for inference, only indexing